fix: move fromBuf functions to a separate file
