+++ ast/fbuffers.go
@@ -2,10 +2,408 @@ package ast
 
 import (
 	"fmt"
+	"regexp"
+	"time"
 
 	flatbuffers "github.com/google/flatbuffers/go"
 	"github.com/influxdata/flux/ast/internal/fbast"
 )
+
+func (p *Position) FromBuf(buf *fbast.Position) {
+	p.Line = int(buf.Line())
+	p.Column = int(buf.Column())
+}
+
+func (l SourceLocation) FromBuf(buf *fbast.SourceLocation) *SourceLocation {
+	l.File = string(buf.File())
+	l.Start.FromBuf(buf.Start(nil))
+	l.End.FromBuf(buf.End(nil))
+	l.Source = string(buf.Source())
+	return &l
+}
+
+func (b *BaseNode) FromBuf(buf *fbast.BaseNode) {
+	b.Loc = SourceLocation{}.FromBuf(buf.Loc(nil))
+	if !b.Loc.IsValid() {
+		b.Loc = nil
+	}
+	if buf.ErrorsLength() != 0 {
+		b.Errors = make([]Error, buf.ErrorsLength())
+		for i := 0; i < buf.ErrorsLength(); i++ {
+			b.Errors[i] = Error{string(buf.Errors(i))}
+		}
+	}
+}
+
+func (p Package) FromBuf(buf []byte) *Package {
+	fbp := fbast.GetRootAsPackage(buf, 0)
+	p.BaseNode.FromBuf(fbp.BaseNode(nil))
+	p.Path = string(fbp.Path())
+	p.Package = string(fbp.Package())
+	p.Files = make([]*File, fbp.FilesLength())
+	for i := 0; i < fbp.FilesLength(); i++ {
+		fbf := new(fbast.File)
+		if !fbp.Files(fbf, i) {
+			p.BaseNode.Errors = append(p.BaseNode.Errors,
+				Error{fmt.Sprintf("Encountered error in deserializing Package.Files[%d]", i)})
+		} else {
+			p.Files[i] = File{}.FromBuf(fbf)
+		}
+	}
+	return &p
+}
+
+func (f File) FromBuf(buf *fbast.File) *File {
+	f.BaseNode.FromBuf(buf.BaseNode(nil))
+	f.Name = string(buf.Name())
+	f.Package = PackageClause{}.FromBuf(buf.Package(nil))
+	if buf.ImportsLength() > 0 {
+		f.Imports = make([]*ImportDeclaration, buf.ImportsLength())
+		for i := 0; i < buf.ImportsLength(); i++ {
+			fbd := new(fbast.ImportDeclaration)
+			if !buf.Imports(fbd, i) {
+				f.BaseNode.Errors = append(f.BaseNode.Errors,
+					Error{fmt.Sprintf("Encountered error in deserializing File.Imports[%d]", i)})
+			} else {
+				f.Imports[i] = ImportDeclaration{}.FromBuf(fbd)
+			}
+		}
+	}
+	if buf.BodyLength() > 0 {
+		var err []Error
+		f.Body, err = statementArrayFromBuf(buf.BodyLength(), buf.Body, "File.Body")
+		if len(err) > 0 {
+			f.BaseNode.Errors = append(f.BaseNode.Errors, err...)
+		}
+	}
+	return &f
+}
+
+func (c PackageClause) FromBuf(buf *fbast.PackageClause) *PackageClause {
+	if buf == nil {
+		return nil
+	}
+	c.BaseNode.FromBuf(buf.BaseNode(nil))
+	c.Name = Identifier{}.FromBuf(buf.Name(nil))
+	return &c
+}
+
+func (d ImportDeclaration) FromBuf(buf *fbast.ImportDeclaration) *ImportDeclaration {
+	d.BaseNode.FromBuf(buf.BaseNode(nil))
+	d.As = Identifier{}.FromBuf(buf.As(nil))
+	d.Path = StringLiteral{}.FromBuf(buf.Path(nil))
+	return &d
+}
+
+func (s Block) FromBuf(buf *fbast.Block) *Block {
+	s.BaseNode.FromBuf(buf.BaseNode(nil))
+	var err []Error
+	s.Body, err = statementArrayFromBuf(buf.BodyLength(), buf.Body, "Block.Body")
+	if len(err) > 0 {
+		s.BaseNode.Errors = append(s.BaseNode.Errors, err...)
+	}
+	return &s
+}
+
+func (s BadStatement) FromBuf(buf *fbast.BadStatement) *BadStatement {
+	s.BaseNode.FromBuf(buf.BaseNode(nil))
+	s.Text = string(buf.Text())
+	return &s
+}
+
+func (s ExpressionStatement) FromBuf(buf *fbast.ExpressionStatement) *ExpressionStatement {
+	s.BaseNode.FromBuf(buf.BaseNode(nil))
+	s.Expression = exprFromBuf("ExpressionStatement.Expression", s.BaseNode, buf.Expression, buf.ExpressionType())
+	return &s
+}
+
+func (s ReturnStatement) FromBuf(buf *fbast.ReturnStatement) *ReturnStatement {
+	s.BaseNode.FromBuf(buf.BaseNode(nil))
+	s.Argument = exprFromBuf("ReturnStatement.Argument", s.BaseNode, buf.Argument, buf.ArgumentType())
+	return &s
+}
+
+func (s OptionStatement) FromBuf(buf *fbast.OptionStatement) *OptionStatement {
+	s.BaseNode.FromBuf(buf.BaseNode(nil))
+	s.Assignment = assignmentFromBuf("OptionStatement.Assignment", s.BaseNode, buf.Assignment, buf.AssignmentType())
+	return &s
+}
+
+func (s BuiltinStatement) FromBuf(buf *fbast.BuiltinStatement) *BuiltinStatement {
+	s.BaseNode.FromBuf(buf.BaseNode(nil))
+	s.ID = Identifier{}.FromBuf(buf.Id(nil))
+	return &s
+}
+
+func (s TestStatement) FromBuf(buf *fbast.TestStatement) *TestStatement {
+	s.BaseNode.FromBuf(buf.BaseNode(nil))
+	s.Assignment = assignmentFromBuf("TestStatement.Assignment",
+		s.BaseNode, buf.Assignment, buf.AssignmentType()).(*VariableAssignment)
+	return &s
+}
+
+func (d VariableAssignment) FromBuf(buf *fbast.VariableAssignment) *VariableAssignment {
+	d.BaseNode.FromBuf(buf.BaseNode(nil))
+	d.ID = Identifier{}.FromBuf(buf.Id(nil))
+	d.Init = exprFromBuf("VariableAssignment.Init", d.BaseNode, buf.Init_, buf.Init_type())
+	return &d
+}
+
+func (a MemberAssignment) FromBuf(buf *fbast.MemberAssignment) *MemberAssignment {
+	a.BaseNode.FromBuf(buf.BaseNode(nil))
+	a.Member = MemberExpression{}.FromBuf(buf.Member(nil))
+	a.Init = exprFromBuf("MemberAssignment.Init", a.BaseNode, buf.Init_, buf.Init_type())
+	return &a
+}
+
+func (e StringExpression) FromBuf(buf *fbast.StringExpression) *StringExpression {
+	e.BaseNode.FromBuf(buf.BaseNode(nil))
+	if buf.PartsLength() > 0 {
+		e.Parts = make([]StringExpressionPart, buf.PartsLength())
+		for i := 0; i < buf.PartsLength(); i++ {
+			fbp := new(fbast.StringExpressionPart)
+			if !buf.Parts(fbp, i) {
+				e.BaseNode.Errors = append(e.BaseNode.Errors,
+					Error{fmt.Sprintf("Encountered error in deserializing StringExpression.Parts[%d]", i)})
+			} else if fbp.TextValue() != nil {
+				e.Parts[i] = TextPart{}.FromBuf(fbp)
+			} else {
+				e.Parts[i] = InterpolatedPart{}.FromBuf(fbp)
+			}
+		}
+	}
+	return &e
+}
+
+func (p TextPart) FromBuf(buf *fbast.StringExpressionPart) *TextPart {
+	p.BaseNode.FromBuf(buf.BaseNode(nil))
+	p.Value = string(buf.TextValue())
+	return &p
+}
+
+func (p InterpolatedPart) FromBuf(buf *fbast.StringExpressionPart) *InterpolatedPart {
+	p.BaseNode.FromBuf(buf.BaseNode(nil))
+	p.Expression = exprFromBuf("InterpolatedPart.Expression", p.BaseNode,
+		buf.InterpolatedExpression, buf.InterpolatedExpressionType())
+	return &p
+}
+
+func (e ParenExpression) FromBuf(buf *fbast.ParenExpression) *ParenExpression {
+	e.BaseNode.FromBuf(buf.BaseNode(nil))
+	e.Expression = exprFromBuf("ParenExpression.Expression", e.BaseNode, buf.Expression, buf.ExpressionType())
+	return &e
+}
+
+func (e CallExpression) FromBuf(buf *fbast.CallExpression) *CallExpression {
+	e.BaseNode.FromBuf(buf.BaseNode(nil))
+	e.Callee = exprFromBuf("CallExpression.Callee", e.BaseNode, buf.Callee, buf.CalleeType())
+	arg := buf.Arguments(nil)
+	if arg != nil {
+		e.Arguments = []Expression{ObjectExpression{}.FromBuf(arg)}
+	}
+	return &e
+}
+
+func (e PipeExpression) FromBuf(buf *fbast.PipeExpression) *PipeExpression {
+	e.BaseNode.FromBuf(buf.BaseNode(nil))
+	e.Argument = exprFromBuf("PipeExpression.Argument", e.BaseNode, buf.Argument, buf.ArgumentType())
+	e.Call = CallExpression{}.FromBuf(buf.Call(nil))
+	return &e
+}
+
+func (e MemberExpression) FromBuf(buf *fbast.MemberExpression) *MemberExpression {
+	e.BaseNode.FromBuf(buf.BaseNode(nil))
+	e.Object = exprFromBuf("MemberExpression.Object", e.BaseNode, buf.Object, buf.ObjectType())
+	e.Property = propertyKeyFromBuf("MemberExpression.Property", e.BaseNode, buf.Property, buf.PropertyType())
+	return &e
+}
+
+func (e IndexExpression) FromBuf(buf *fbast.IndexExpression) *IndexExpression {
+	e.BaseNode.FromBuf(buf.BaseNode(nil))
+	e.Array = exprFromBuf("IndexExpression.Array", e.BaseNode, buf.Array, buf.ArrayType())
+	e.Index = exprFromBuf("IndexExpression.Index", e.BaseNode, buf.Index, buf.IndexType())
+	return &e
+}
+
+func (e FunctionExpression) FromBuf(buf *fbast.FunctionExpression) *FunctionExpression {
+	e.BaseNode.FromBuf(buf.BaseNode(nil))
+	if buf.ParamsLength() > 0 {
+		e.Params = make([]*Property, buf.ParamsLength())
+		for i := 0; i < buf.ParamsLength(); i++ {
+			fbp := new(fbast.Property)
+			if !buf.Params(fbp, i) {
+				e.BaseNode.Errors = append(e.BaseNode.Errors,
+					Error{fmt.Sprintf("Encountered error in deserializing FunctionExpression.Params[%d]", i)})
+			} else {
+				e.Params[i] = Property{}.FromBuf(fbp)
+			}
+		}
+	}
+	t := new(flatbuffers.Table)
+	if buf.Body(t) {
+		switch buf.BodyType() {
+		case fbast.ExpressionOrBlockBlock:
+			b := new(fbast.Block)
+			b.Init(t.Bytes, t.Pos)
+			e.Body = Block{}.FromBuf(b)
+		case fbast.ExpressionOrBlockWrappedExpression:
+			we := new(fbast.WrappedExpression)
+			we.Init(t.Bytes, t.Pos)
+			e.Body = exprFromBuf("FunctionExpression.Body", e.BaseNode, we.Expr, we.ExprType())
+		default:
+			e.BaseNode.Errors = append(e.BaseNode.Errors,
+				Error{"Encountered error in deserializing FunctionExpression.Body"})
+		}
+	}
+	return &e
+}
+
+func (e BinaryExpression) FromBuf(buf *fbast.BinaryExpression) *BinaryExpression {
+	e.BaseNode.FromBuf(buf.BaseNode(nil))
+	e.Operator = opMap[buf.Operator()]
+	e.Left = exprFromBuf("BinaryExpression.Left", e.BaseNode, buf.Left, buf.LeftType())
+	e.Right = exprFromBuf("BinaryExpression.Right", e.BaseNode, buf.Right, buf.RightType())
+	return &e
+}
+
+func (e UnaryExpression) FromBuf(buf *fbast.UnaryExpression) *UnaryExpression {
+	e.BaseNode.FromBuf(buf.BaseNode(nil))
+	e.Operator = opMap[buf.Operator()]
+	e.Argument = exprFromBuf("UnaryExpression.Argument", e.BaseNode, buf.Argument, buf.ArgumentType())
+	return &e
+}
+
+func (e LogicalExpression) FromBuf(buf *fbast.LogicalExpression) *LogicalExpression {
+	e.BaseNode.FromBuf(buf.BaseNode(nil))
+	e.Operator = logOpMap[buf.Operator()]
+	e.Left = exprFromBuf("LogicalExpression.Left", e.BaseNode, buf.Left, buf.LeftType())
+	e.Right = exprFromBuf("LogicalExpression.Right", e.BaseNode, buf.Right, buf.RightType())
+	return &e
+}
+
+func (e ArrayExpression) FromBuf(buf *fbast.ArrayExpression) *ArrayExpression {
+	e.BaseNode.FromBuf(buf.BaseNode(nil))
+	var err []Error
+	e.Elements, err = exprArrayFromBuf(buf.ElementsLength(), buf.Elements, "ArrayExpression.Elements")
+	if len(err) > 0 {
+		e.BaseNode.Errors = append(e.BaseNode.Errors, err...)
+	}
+	return &e
+}
+
+func (e ObjectExpression) FromBuf(buf *fbast.ObjectExpression) *ObjectExpression {
+	e.BaseNode.FromBuf(buf.BaseNode(nil))
+	e.With = Identifier{}.FromBuf(buf.With(nil))
+	e.Properties = make([]*Property, buf.PropertiesLength())
+	for i := 0; i < buf.PropertiesLength(); i++ {
+		fbp := new(fbast.Property)
+		if !buf.Properties(fbp, i) {
+			e.BaseNode.Errors = append(e.BaseNode.Errors,
+				Error{fmt.Sprintf("Encountered error in deserializing ObjectExpression.Properties[%d]", i)})
+		} else {
+			e.Properties[i] = Property{}.FromBuf(fbp)
+		}
+	}
+	return &e
+}
+
+func (e ConditionalExpression) FromBuf(buf *fbast.ConditionalExpression) *ConditionalExpression {
+	e.BaseNode.FromBuf(buf.BaseNode(nil))
+	e.Test = exprFromBuf("ConditionalExpression.Test", e.BaseNode, buf.Test, buf.TestType())
+	e.Consequent = exprFromBuf("ConditionalExpression.Consequent", e.BaseNode, buf.Consequent, buf.ConsequentType())
+	e.Alternate = exprFromBuf("ConditionalExpression.Alternate", e.BaseNode, buf.Alternate, buf.AlternateType())
+	return &e
+}
+
+func (p Property) FromBuf(buf *fbast.Property) *Property {
+	p.BaseNode.FromBuf(buf.BaseNode(nil))
+	// deserialize key
+	p.Key = propertyKeyFromBuf("Property.Key", p.BaseNode, buf.Key, buf.KeyType())
+	// deserialize value
+	p.Value = exprFromBuf("Property.Value", p.BaseNode, buf.Value, buf.ValueType())
+	return &p
+}
+
+func (i Identifier) FromBuf(buf *fbast.Identifier) *Identifier {
+	if buf == nil {
+		return nil
+	}
+	i.BaseNode.FromBuf(buf.BaseNode(nil))
+	i.Name = string(buf.Name())
+	return &i
+}
+
+func (p PipeLiteral) FromBuf(buf *fbast.PipeLiteral) *PipeLiteral {
+	p.BaseNode.FromBuf(buf.BaseNode(nil))
+	return &p
+}
+
+func (l StringLiteral) FromBuf(buf *fbast.StringLiteral) *StringLiteral {
+	l.BaseNode.FromBuf(buf.BaseNode(nil))
+	l.Value = string(buf.Value())
+	return &l
+}
+
+func (l BooleanLiteral) FromBuf(buf *fbast.BooleanLiteral) *BooleanLiteral {
+	l.BaseNode.FromBuf(buf.BaseNode(nil))
+	l.Value = buf.Value()
+	return &l
+}
+
+func (l FloatLiteral) FromBuf(buf *fbast.FloatLiteral) *FloatLiteral {
+	l.BaseNode.FromBuf(buf.BaseNode(nil))
+	l.Value = buf.Value()
+	return &l
+}
+
+func (l IntegerLiteral) FromBuf(buf *fbast.IntegerLiteral) *IntegerLiteral {
+	l.BaseNode.FromBuf(buf.BaseNode(nil))
+	l.Value = buf.Value()
+	return &l
+}
+
+func (l UnsignedIntegerLiteral) FromBuf(buf *fbast.UnsignedIntegerLiteral) *UnsignedIntegerLiteral {
+	l.BaseNode.FromBuf(buf.BaseNode(nil))
+	l.Value = buf.Value()
+	return &l
+}
+
+func (l RegexpLiteral) FromBuf(buf *fbast.RegexpLiteral) *RegexpLiteral {
+	l.BaseNode.FromBuf(buf.BaseNode(nil))
+	var err error
+	if l.Value, err = regexp.Compile(string(buf.Value())); err != nil {
+		l.BaseNode.Errors = append(l.BaseNode.Errors,
+			Error{fmt.Sprintf("Encountered error in deserializing RegexpLiteral.Values: %s", err.Error())})
+	}
+	return &l
+}
+
+func (d Duration) FromBuf(buf *fbast.Duration) Duration {
+	d.Magnitude = buf.Magnitude()
+	d.Unit = fbast.EnumNamesTimeUnit[buf.Unit()]
+	return d
+}
+
+func (l DurationLiteral) FromBuf(buf *fbast.DurationLiteral) *DurationLiteral {
+	l.BaseNode.FromBuf(buf.BaseNode(nil))
+	l.Values = make([]Duration, buf.ValuesLength())
+	for i := 0; i < buf.ValuesLength(); i++ {
+		d := new(fbast.Duration)
+		if !buf.Values(d, i) {
+			l.BaseNode.Errors = append(l.BaseNode.Errors,
+				Error{fmt.Sprintf("Encountered error in deserializing DurationLiteral.Values[%d]", i)})
+		} else {
+			l.Values[i] = Duration{}.FromBuf(d)
+		}
+	}
+	return &l
+}
+
+func (l DateTimeLiteral) FromBuf(buf *fbast.DateTimeLiteral) *DateTimeLiteral {
+	l.BaseNode.FromBuf(buf.BaseNode(nil))
+	l.Value = time.Unix(buf.Secs(), int64(buf.Nsecs())).In(time.FixedZone("DateTimeLiteral offset", int(buf.Offset())))
+	return &l
+}
 
 type stmtGetterFn func(obj *fbast.WrappedStatement, j int) bool
